requirements.txt
pandas
numpy
pyarrow
pyyaml
bottleneck
numexpr
flask
flask-cors
scipy
numba
